Added implementations for validation of simple-move
Added the logic to check for simple moves in the SimpleMove class.
Overrided the Move class' executeMove() and validateMove() methods
in order for the Move to validate itself based on the rules of the
move.

NOTE: Did not implement any tests and did not implement PostValidate
Move yet.

Changed:
- SimpleMove: Override of superclass Move methods as stated above
- Move: changed the signature of execute and validate move because the
moves need to know who is making the move in order to propery validate.
- Board: Updated the move() function to remove checks since moves will
do it themselves and updated the signature to use Positions.
Update the space is valid to use positions.