Removed reliance on TurnControllers movesMade attr
Since the TurnController is invoked every single time a handle method is
called, then that means the moves made does not persist between calls
and the movesMade does not actually add any functionality. Therefore, I
made it rely on the size of the "Queue" of moves in the Game Class. This
way there is a persistent number of moves across calls.  Also, removed
the function returnMessageAndResetMoves because it relied on the attr i
deleted