Added multi-move changes to JumpMove class
When checking for a jump move, the code would only check if there is a
piece on the starting space of a move. However, if the turn is a
multi-jump-move, the moves will yet to be applied to the model. To get
around this I added a check that it checks the current turn's moves list
and sees if the last move made ends where the current move starts