Changed implemntation of move
Since there are many errors that can come along with incorrectly moving
a piece, the error message should be set by the move itself and not be
decided after the fact. However, our code does not reflect this. I added
a new attribute to Move super class that allows the moves to store their
own error message. Then they will be able to pass this to the turn
controller