Added methods to start and end resignation process
In order to resign a pleyr, based on our current class interactions, the
resignation process must be split up into two parts.
1.) a player leaves the game
2.) then the rest of the data is deleted

This is broken up into two parts because GetGameRoute needs to
differentiate between players that have left a game and a player that do
not share a common game. By allowing access to the data after one person
has left, this will let the server user the game until the other player
has recieved the message that the other player has left