Fix race condition with ending the game
When a player submits their turn and wins the game, if they were really
quick to click "ok" on the "you won!" alert window, they would be
redirected back to the game when they go to the home page. This is
because the webserver has not yet changed the game state, so it looks
like they are still in a game.

The fix is really dirty. I added a 3-second wait to the javascript so
they are forced to wait for the webserver to update the game state
before redirecting back to the home page.

I'm so sorry.